requirements.txt
numpy
pandas
tqdm
transformers
sentence-transformers
accelerate
bitsandbytes
faiss-cpu
sqlite-utils
tavily-python
serpapi
python-dotenv
fastapi
uvicorn
requests
langchain
langchain-community
duckduckgo-search
torch
arxiv  
